pyproject.toml
[tool.poetry]
name = "gitlab-actions"
version = "0.1.0"
description = "GitLab actions"
authors = ["Hirofumi Tsuzuki <hrfm.tsuzuki@gmail.com>"]
readme = "README.md"
repository = "https://github.com/hrfmtzk/gitlab-actions"

[tool.poetry.dependencies]
python = "^3.9"
python-gitlab = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.6.4"
mypy = "^0.790"
bandit = "^1.6.2"
xenon = "^0.7.1"

[tool.black]
line-length = 80

[tool.isort]
include_trailing_comma = true
line_length = 80
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
